+++ src/main/java/com/acs560/cricket_analyzer/Repository/CountriesRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import com.acs560.cricket_analyzer.entities.CountryEntity;
 
 
-public interface CountriesRepository extends CrudRepository<CountryEntity, Integer> {
+public interface CountriesRepository extends CrudRepository<CountryEntity, Long> {
 
 	CountryEntity findByTeam(String team);
