+++ src/main/java/com/acs560/cricket_analyzer/Services/PlayerServices.java
@@ -8,15 +8,13 @@ import jakarta.validation.Valid;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.stereotype.Service;
 
-//@Service
 public interface PlayerServices {
 	
-	Optional<Player> getPlayers(String name, int countryId);
+	Optional<Player> getPlayers(String name, Long countryId);
 	List<Player> getPlayers();
-	List<Player> getPlayersByCountry(int countryId);
-	List<Player> getPlayersByCountryAndNotouts(int countryId, int notouts);
+	List<Player> getPlayersByCountry(Long countryId);
+	List<Player> getPlayersByCountryAndNotouts(Long countryId, int notouts);
 
 	Player add(@Valid Player player);
 	boolean delete(@Valid Player player);
